requirements.txt
# Core ML/AI
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
torchmetrics>=0.11.0
timm>=0.9.0
segmentation-models-pytorch>=0.3.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0
albumentations>=1.3.0

# 3D Processing & Geometry
open3d>=0.17.0
trimesh>=3.23.0
pyvista>=0.42.0
meshio>=5.3.0
pygltflib>=1.15.0

# Blender Integration
bpy>=3.6.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Configuration & Validation
pydantic>=2.4.0
pyyaml>=6.0.0
jsonschema>=4.19.0

# Utilities
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
wandb>=0.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
pre-commit>=3.4.0

# Optional: Advanced 3D
# kaolin>=0.15.0  # NVIDIA's 3D ML library
# pytorch3d>=0.7.0  # Facebook's 3D ML library
